schoolmanagement/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
distlib==0.3.8
Django==4.2.17
django-allauth==64.2.1
django-cors-headers==4.4.0
django-filter==24.3
django-oauth-toolkit==3.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
filelock==3.16.0
idna==3.8
jwcrypto==1.5.6
oauthlib==3.2.2
phonenumbers==8.13.45
pillow==10.4.0
platformdirs==4.3.3
pycountry==24.6.1
pycparser==2.22
PyJWT==2.9.0
requests==2.32.3
sqlparse==0.5.1
twilio==9.3.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
virtualenv==20.26.4
virtualenvwrapper-win==1.2.7